requirements.txt
Flask==3.0.0
pandas==2.2.0
anthropic==0.21.0
gunicorn==21.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
